backend/app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
chromadb==0.4.0
llama-cpp-python==0.1.76  # optional; if install fails we'll use subprocess fallback
python-dotenv==1.0.0
pydantic==1.10.9

